refactor to handle startup race condition
