+++ alerting/src/main/proto/plugin.proto
@@ -29,7 +29,7 @@ syntax = "proto3";
 
 package plugin;
 option java_multiple_files = true;
-option java_package = "org.opennms.poc.plugin.config";
+option java_package = "org.opennms.poc.alerting.proto";
 
 message PluginConfigMessage {
   message PluginConfigMeta {
