+++ alerting/src/main/resources/OSGI-INF/blueprint/alerting.xml
@@ -4,19 +4,26 @@
   xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
   xsi:schemaLocation="http://www.osgi.org/xmlns/blueprint/v1.0.0 http://www.osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd">
 
-  <cm:property-placeholder persistent-id="org.opennms.poc.ignite.worker" update-strategy="reload">
+  <cm:property-placeholder persistent-id="org.opennms.poc.alerting" update-strategy="reload">
     <cm:default-properties>
       <cm:property name="routing.uri" value="seda:registration"/>
     </cm:default-properties>
   </cm:property-placeholder>
 
-  <camel:camelContext id="minionRoutingContext" allowUseOriginalMessage="false">
+  <camel:camelContext id="alertRoutingContext" allowUseOriginalMessage="false">
     <camel:contextScan/>
     <camel:template id="registrationProducer" defaultEndpoint="{{routing.uri}}"/>
   </camel:camelContext>
 
+  <reference id="messageDispatcherFactory" interface="org.opennms.horizon.ipc.sink.api.MessageDispatcherFactory" />
+
   <bean id="alertingService" class = "org.opennms.poc.alerting.AlertingServiceImpl">
     <argument ref="registrationProducer"/>
+  </bean>
+
+  <bean id="minionRouting" class="org.opennms.poc.alerting.MinionRouting">
+    <argument value="${routing.uri}"/>
+    <argument ref="messageDispatcherFactory"/>
   </bean>
 
   <service ref="alertingService" interface="org.opennms.poc.alerting.AlertingService"/>
