+++ model/src/main/java/org/opennms/poc/ignite/model/workflows/PluginMetadata.java
@@ -5,7 +5,6 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
-import org.opennms.poc.ignite.model.workflows.WorkflowType;
 import org.opennms.poc.plugin.api.FieldConfigMeta;
 
 @AllArgsConstructor
